simpler structure: remove eval & dataset dirs
