--- tests/utils.py
@@ -1,6 +1,0 @@
-import random
-import string
-
-
-def random_string(length: int) -> str:
-    return "".join(random.choices(string.ascii_lowercase, k=length))
